Basic Nav code V 2
accepts outside dimensions of box to be mowed as variables in code. has option for course correction 
Intended to test mower maneuvering ability. 
Modified from initial version to account for incorrect behavior. Includes updated turn and stop functions primarily. as well as option for course correction 
NOT TESTED ON HARDWARE